Update automator action buttons
